cpu/stm32: introduce STM32MP1_ENGINEERING_MODE flag
In Engineering mode (BOOT0 off and BOOT2 on), only the Cortex-M4
core is running. It means that all clocks have to be setup
by the Cortex-M4 core.
In other modes, the clocks are setup by the Cortex-A7 and then should
not be setup by Cortex-M4.
STM32MP1_ENGINEERING_MODE macro have to be defined in Engineering mode
to ensure clocks configuration.
This macro can also be used in peripheral drivers to define clock source
for each one.

Signed-off-by: Gilles DOFFE <gilles.doffe@savoirfairelinux.com>